TTK-22272 No validate Opencast host and paths using env vars conf
Conf like:
```
pumukit_opencast:
    host: "%env(PROP_ORG_OPENCASTPROJECT_ADMIN_UI_URL)%"
    username: "%env(ORG_OPENCASTPROJECT_SECURITY_DIGEST_USER)%" # digest username
    password: '%env(ORG_OPENCASTPROJECT_SECURITY_DIGEST_PASSWORD)%' # digest password
    player: /engage/ui/watch.html
    use_redirect: false
    batchimport_inverted: false
...
    url_mapping:
        - url: 'https://"%env(PROP_ORG_OPENCASTPROJECT_ENGAGE_UI_URL)%"/static/mh_default_org/engage-player'
          path: /mnt/opencast/downloads/mh_default_org/engage-player

```

If the value of the configuration is an environment variable as
'%env(PUMUKIT_OC_HOST)%', the extension doesn't know the correct value when
the Symfony cache is generated.

Curernt solution skip the validation.

PumukitOpencastExtension is messy, the future solution is to create a separate
package with the OpencastBundle and to use a typical xml file to define the services.
To validations and initial configurations the premmium solution should be create
tagged services and commands.